+++ packages/sdk/jest.config.js
@@ -1,0 +1,12 @@
+module.exports = {
+  preset: 'ts-jest',
+  testEnvironment: 'node',
+  collectCoverage: true,
+  coverageDirectory: 'coverage',
+  verbose: true,
+  globals: {
+    'ts-jest': {
+      isolatedModules: true,
+    },
+  },
+};